// src/modules/notifications/domain/enum/notification-type.enum.ts
export enum NotificationType {
  EVENT_PUBLISHED = 'EVENT_PUBLISHED',
  EVENT_CANCELLED = 'EVENT_CANCELLED',
  EVENT_UPDATED = 'EVENT_UPDATED',
  EVENT_REMINDER = 'EVENT_REMINDER',
  EVENT_JOINED = 'EVENT_JOINED',
  EVENT_LEFT = 'EVENT_LEFT',
}

export enum NotificationStatus {
  PENDING = 'PENDING',
  SENT = 'SENT',
  READ = 'READ',
  FAILED = 'FAILED',
}